initial commit inside clion.
changed the environment of the project to clion.
