+++ src/head.h
@@ -1,4 +1,6 @@
 #include <iostream>
 #include <stdlib.h>
+#include <string.h>
+#include <cmath>
 
 using namespace std;
